feat(view): profile image 가져오는 함수 공용화
