+++ packages/view/src/components/Statistics/AuthorBarChart/AuthorBarChart.tsx
@@ -2,22 +2,18 @@ import type { ChangeEvent, MouseEvent } from "react";
 import { useRef, useEffect, useState } from "react";
 import * as d3 from "d3";
 
-import type { ClusterNode } from "types";
+import type { ClusterNode, SrcInfo } from "types";
 import { useGlobalData } from "hooks";
+import { getAuthorProfileImgSrc } from "utils/author";
 
 import { useGetSelectedData } from "../Statistics.hook";
 
-import type {
-  AuthorDataType,
-  MetricType,
-  SrcInfo,
-} from "./AuthorBarChart.type";
+import type { AuthorDataType, MetricType } from "./AuthorBarChart.type";
 import {
   convertNumberFormat,
   getDataByAuthor,
   sortDataByAuthor,
   sortDataByName,
-  getAuthorProfileImgSrc,
 } from "./AuthorBarChart.util";
 import { DIMENSIONS, METRIC_TYPE } from "./AuthorBarChart.const";
