+++ packages/view/src/components/VerticalClusterList/Summary/Summary.util.ts
@@ -1,14 +1,12 @@
-import md5 from "md5";
-
 import type {
   GlobalProps,
   CommitNode,
   ClusterNode,
   SelectedDataProps,
 } from "types";
+import { getAuthorProfileImgSrc } from "utils/author";
 
-import { GITHUB_URL, GRAVATA_URL } from "./Summary.const";
-import type { AuthSrcMap, Cluster, SrcInfo } from "./Summary.type";
+import type { AuthSrcMap, Cluster } from "./Summary.type";
 
 export function getInitData(data: GlobalProps["data"]) {
   const clusters: Cluster[] = [];
@@ -79,27 +77,6 @@ function getAuthorNames(data: ClusterNode[]) {
     .flat();
   const setAuthorNames = new Set(authorNames);
   return Array.from(setAuthorNames);
-}
-
-function getAuthorProfileImgSrc(authorName: string): Promise<SrcInfo> {
-  return new Promise((resolve) => {
-    const img = new Image();
-
-    img.onload = () => {
-      const { src } = img;
-      const srcInfo: SrcInfo = {
-        key: authorName,
-        value: src,
-      };
-      resolve(srcInfo);
-    };
-
-    img.onerror = () => {
-      img.src = `${GRAVATA_URL}/${md5(authorName)}}?d=identicon&f=y`;
-    };
-
-    img.src = `${GITHUB_URL}/${authorName}.png?size=30`;
-  });
 }
 
 export async function getAuthSrcMap(data: ClusterNode[]) {
